Extract pwm of intervals from a motif database

Extract pwm of intervals from a motif database

Usage

gextract_pwm(
  intervals,
  motifs = NULL,
  dataset = all_motif_datasets(),
  spat = NULL,
  spat_min = 0,
  spat_max = NULL,
  bidirect = TRUE,
  prior = 0,
  parallel = getOption("prego.parallel", TRUE)
)

Arguments

intervals

misha intervals set

motifs

names of specific motifs to extract from the dataset

dataset

a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example HOMER_motifs or JASPAR_motifs, or all_motif_datasets().

spat

a data frame with the spatial model (as returned from the $spat slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.

spat_min

start of the spatial model from the beginning of the sequence (in bp)

spat_max

end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model would end at the end of the sequence.

bidirect

is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.

prior

a prior probability for each nucleotide.

parallel

logical, whether to use parallel processing

Value

The intervals set with additional columns per motif, containing the pwm of each interval for each motif

Examples

# \dontrun{
library(misha)
gdb.init_examples()
pwms <- gextract_pwm("annotations")
pwms[, 1:20]
#>   chrom start   end strand  remark HOMER.AARE HOMER.ABF1 HOMER.AMYB HOMER.AP_1
#> 1  chr1    20  2000      1 remarkA -12.989733  -12.25663  -6.532580 -10.476537
#> 2  chr2    20  2000      1 remarkB -11.239439  -11.09528  -7.208153 -11.184667
#> 3  chr2  3000  8000     -1 remarkC -10.006383  -10.59158  -6.676258  -4.385600
#> 4  chr2  9000 11000      1 remarkD  -5.368375  -11.75360  -9.622133  -7.448188
#>   HOMER.AP_2alpha HOMER.AP_2gamma HOMER.ARE HOMER.AR_halfsite HOMER.Ap4
#> 1      -11.174854      -11.531642 -18.29645         -3.582500 -5.594416
#> 2       -8.121261       -7.816772 -16.74446         -5.461981 -4.898031
#> 3      -11.295572      -10.678673 -14.98829         -6.295091 -4.292560
#> 4      -11.155062      -11.075082 -15.17114         -6.004141 -3.999556
#>   HOMER.ArntAhr HOMER.Ascl1 HOMER.Atf1 HOMER.Atf2 HOMER.Atf3 HOMER.Atf4
#> 1     -5.997535   -7.424507  -6.736964  -9.205512 -14.353178 -12.614318
#> 2     -5.877813   -7.202877  -7.836501  -9.936964 -14.161862 -11.135440
#> 3     -6.492581   -7.565849  -5.986350  -8.801591  -6.384101  -9.361985
#> 4     -5.802579   -7.054382  -5.009895  -7.242763 -10.818832  -3.324120
# }