Run PWM regression on clusters

Run PWM regression on clusters

Usage

regress_pwm_clusters(sequences, clusters, two_phase = TRUE, ...)

Arguments

sequences

A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through toupper)

clusters

a vector with the cluster assignments for each sequence

...

Arguments passed on to regress_pwm, regress_pwm_two_phase

response

A matrix of response variables - number of rows should equal the number of sequences

motif

Initial motif to start the regression from. Can be either a string with a kmer where the character "*" indicates a wildcard or a data frame with a pre-computed PSSM (see thre slot pssm in the return value of this function). If NULL - a K-mer screen would be performed in order to find the best kmer for initialization.

motif_length

Length of the seed motif. If the motif is shorter than this, it will be extended by wildcards (stars). Note that If the motif is longer than this, it will not be truncated.

score_metric

metric to use for optimizing the PWM. One of "r2" or "ks". When using "ks" the response variable should be a single vector of 0 and 1.

bidirect

is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.

spat_min

start of the spatial model from the beginning of the sequence (in bp)

spat_max

end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model would end at the end of the sequence.

spat_bin

size of the spatial bin (in bp).

improve_epsilon

minimum improve in the objective function to continue the optimization

min_nuc_prob

minimum nucleotide probability in every iteration

unif_prior

uniform prior for nucleotide probabilities

include_response

include the response in the resulting list (default: TRUE)

verbose

show verbose messages.

seed

random seed

consensus_single_thresh,consensus_double_thresh

thresholds for the consensus sequence calculation (single and double nucleotides)

kmer_length

The number of non-gap characters in motifs that will be screened

is_train

a boolean vector that determine which subset of sequences to use when screening

two_phase_sample_frac

fraction of the dataset to sample for the first phase of the optimization (default: 0.1). When response is categorical (0 and 1), the sampling would be stratified by the category, i.e. two_phase_sample_frac can be a vector of length 2 with the fraction of 0 and 1 responses to sample respectively.

first_phase_idxs

indices of the sequences to use in the first phase of the optimization. If NULL, the indices would be sampled using two_phase_sample_frac.

first_phase_metric

metric to use in order to choose the best motif in the first phase of the optimization. One of 'ks' or 'r2'. Note that unlike score_metric which is used in the regression itself, this metric is used only for choosing the best motif in the first phase of the optimization out of all the runs on the sampled dataset.

min_gap,max_gap

the length of a gap to be considered in the pattern. Only one gap, of length min_gap:max_gap, is being used, and is located anywhere in the motif. Note that this greatly expand the search space (and increase multiple tesing severly).