Run a 2-phase PWM regression

The first phase of the optimization would be performed with a sampled dataset of size two_phase_sample_frac where different candidates of kmers would be regressed in order to find the best seed. Thenm the optimization would be performed on the full dataset while initializing the motif from the sampled dataset. You can also give explicit indices of the sequences to use in the first phase using first_phase_idxs.

Usage

regress_pwm_two_phase(
  sequences,
  response,
  motif_length = 15,
  score_metric = "r2",
  bidirect = TRUE,
  spat_min = 0,
  spat_max = NULL,
  spat_bin = 50,
  improve_epsilon = 0.0001,
  min_nuc_prob = 0.001,
  unif_prior = 0.05,
  is_train = NULL,
  include_response = TRUE,
  seed = 60427,
  verbose = FALSE,
  kmer_length = 5:8,
  min_gap = 0,
  max_gap = 1,
  min_kmer_cor = 0.1,
  two_phase_sample_frac = 0.1,
  first_phase_idxs = NULL,
  first_phase_metric = "r2",
  ...
)

Arguments

sequences

A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through toupper)

response

A matrix of response variables - number of rows should equal the number of sequences

motif_length

Length of the seed motif. If the motif is shorter than this, it will be extended by wildcards (stars). Note that If the motif is longer than this, it will not be truncated.

score_metric

metric to use for optimizing the PWM. One of "r2" or "ks". For categorical response variables (0 and 1), "ks" is recommended, while for continuous response variables, "r2" is recommended. Default is "r2". When using "ks" the response variable should be a single vector of 0 and 1.

bidirect

is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.

spat_min

start of the spatial model from the beginning of the sequence (in bp)

spat_max

end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model would end at the end of the sequence.

spat_bin

size of the spatial bin (in bp).

improve_epsilon

minimum improve in the objective function to continue the optimization

min_nuc_prob

minimum nucleotide probability in every iteration

unif_prior

uniform prior for nucleotide probabilities

is_train

a boolean vector that determine which subset of sequences to use when screening

include_response

include the response in the resulting list (default: TRUE)

seed

random seed

verbose

verbosity of the second phase of the optimization.

kmer_length

a vector of kmer lengths to screen in order to find the best seed motif.

two_phase_sample_frac

fraction of the dataset to sample for the first phase of the optimization (default: 0.1). When response is categorical (0 and 1), the sampling would be stratified by the category, i.e. two_phase_sample_frac can be a vector of length 2 with the fraction of 0 and 1 responses to sample respectively.

first_phase_idxs

indices of the sequences to use in the first phase of the optimization. If NULL, the indices would be sampled using two_phase_sample_frac.

first_phase_metric

metric to use in order to choose the best motif in the first phase of the optimization. One of 'ks' or 'r2'. Note that unlike score_metric which is used in the regression itself, this metric is used only for choosing the best motif in the first phase of the optimization out of all the runs on the sampled dataset.

Examples

res <- regress_pwm_two_phase(sequences_example, response_mat_example, two_phase_sample_frac = 0.1)
#> ℹ Performing two phase optimization
#> 
#> ── First phase ──
#> 
#> ℹ Sampling 0.1 of the dataset for the first phase
#> 
#> ── Generate candidate kmers 
#> 
#> ── Regress each candidate kmer on sampled data 
#> ℹ Running regression on 15 candidate kmers
#> → "TGG*TCTT", score: 0.0334954325999651, 0.164550525909299, 0.0896455863787199, 0.0303649822106501, and 0.0371369393142067
#> → "TTATC", score: 0.171306534272033, 0.151836110451098, 0.268797950313114, 0.104559944458828, and 0.121327188109165
#> → "CAGAA", score: 0.134099165696488, 0.0685569160728428, 0.329522809117285, 0.210776665681083, and 0.108384495123907
#> → "CT*AAA", score: 0.205198806721792, 0.221297674795456, 0.233750831694291, 0.0952165386555909, and 0.0853449899694711
#> → "AAAGTT", score: 0.369531065354674, 0.202795723060236, 0.0652501723729291, 0.0494363274925201, and 0.0170479585890799
#> → "GGCTTT", score: 0.000455710095225261, 0.0208364939128581, 0.00280519864103107, 0.421594657973522, and 0.119508027266349
#> → "AAACCTG", score: 0.223661332750854, 0.120312064378447, 0.0189499490996121, 0.0487289202686493, and 0.0224577278278001
#> ■■■■■■■■■■■■■■■                   47% | ETA:  3s
#> 
#> → "ACACAGT", score: 0.310099617648068, 0.129926793347138, 0.237729463407313, 0.0696188258908497, and 0.0548640689472569
#> ■■■■■■■■■■■■■■■                   47% | ETA:  3s
#> 
#> → "ATCTCCT", score: 0.358506251883035, 0.219643750363351, 0.183357513596746, 0.00922714506957671, and 0.00235824433334706
#> ■■■■■■■■■■■■■■■                   47% | ETA:  3s
#> ■■■■■■■■■■■■■■■■■■■               60% | ETA:  2s
#> 
#> → "TTTGG*TC", score: 0.33279330500321, 0.400244115891077, 0.00925390519431678, 0.0558430766365451, and 0.0385623862063643
#> ■■■■■■■■■■■■■■■■■■■               60% | ETA:  2s
#> ■■■■■■■■■■■■■■■■■■■■■             67% | ETA:  2s
#> 
#> → "AGAAAGAG", score: 0.301153707150868, 0.230654299523971, 0.16418397268167, 0.108099889242997, and 0.0638676817635785
#> ■■■■■■■■■■■■■■■■■■■■■             67% | ETA:  2s
#> ■■■■■■■■■■■■■■■■■■■■■■■           73% | ETA:  2s
#> 
#> → "GAGAAAGA", score: 0.205597731614181, 0.223967890793122, 0.0292138289586091, 0.0423770470118631, and 0.0930463459704582
#> ■■■■■■■■■■■■■■■■■■■■■■■           73% | ETA:  2s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         80% | ETA:  1s
#> 
#> → "ATTTG*GTC", score: 0.310075846008932, 0.353914551077865, 0.0205856643489672, 0.171526959518067, and 0.121007785805592
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         80% | ETA:  1s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87% | ETA:  1s
#> 
#> → "GA*AAAGAG", score: 0.354828669632276, 0.14475227911855, 0.0283621407660442, 0.137758544983709, and 0.0743022919663322
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87% | ETA:  1s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     93% | ETA:  0s
#> 
#> → "TGC*ATCCT", score: 0.352744289107446, 0.213821761622386, 0.0421405960558539, 0.104052024128266, and 0.0458817204803453
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     93% | ETA:  0s
#> 
#> Error in res_s_list[[which.max(scores)]]: subscript out of bounds
res$pssm
#> Error in eval(expr, envir, enclos): object 'res' not found
res$spat
#> Error in eval(expr, envir, enclos): object 'res' not found
head(res$pred)
#> Error in head(res$pred): object 'res' not found