Cross-validate a PWM regression model

Perform cross-validation on a PWM regression model. You can either provide explicit folds, or use the nfolds argument to set the number of folds. If the response is binary (0 and 1) or a categories vector is given, the folds would be stratified by the response/categories.

Usage

regress_pwm.cv(
  sequences,
  response,
  nfolds = NULL,
  metric = NULL,
  folds = NULL,
  categories = NULL,
  use_sample = TRUE,
  seed = 60427,
  parallel = getOption("prego.parallel", FALSE),
  add_full_model = TRUE,
  ...
)

Arguments

sequences

A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through toupper)

response

A matrix of response variables - number of rows should equal the number of sequences

nfolds

number of folds for cross-validation. Can be NULL if folds are provided.

metric

metric to use for cross-validation. One of 'ks' or 'r2'. If NULL - 'ks' would be set for binary response and 'r2' for continuous response.

folds

vector of fold numbers for each sequence (optional)

categories

vector of categories for each sequence (optional)

use_sample

whether to use sampled optimization or not.

seed

random seed

parallel

whether to run the cross-validation in parallel.

add_full_model

whether to add the full model (without cross-validation) to the results.

...

Arguments passed on to regress_pwm, regress_pwm.sample

motif

Initial motif to start the regression from. Can be either a string with a kmer where the character "*" indicates a wildcard or a data frame with a pre-computed PSSM (see the slot pssm in the return value of this function). If NULL - a K-mer screen would be performed in order to find the best kmer for initialization. If init_from_dataset is TRUE, the regression would be initialized from the PSSM of the best motif in the dataset.

init_from_dataset

initialize the regression from the PSSM of the best motif in motif_dataset, using final_metric as the metric. If TRUE, the motif parameter would be ignored. See screen_pwm for more details.

motif_length

Length of the seed motif. If the motif is shorter than this, it will be extended by wildcards (stars). Note that If the motif is longer than this, it will not be truncated.

score_metric

metric to use for optimizing the PWM. One of "r2" or "ks". When using "ks" the response variable should be a single vector of 0 and 1.

bidirect

is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.

spat_min

start of the spatial model from the beginning of the sequence (in bp)

spat_max

end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model would end at the end of the sequence.

spat_bin

size of the spatial bin (in bp).

spat_model

a previously computed spatial model (see spat) in the return value of this function. This can only be used when motif is a previously computed PSSM.

improve_epsilon

minimum improve in the objective function to continue the optimization

min_nuc_prob

minimum nucleotide probability in every iteration

unif_prior

uniform prior for nucleotide probabilities

include_response

include the response in the resulting list (default: TRUE)

verbose

show verbose messages.

consensus_single_thresh,consensus_double_thresh

thresholds for the consensus sequence calculation (single and double nucleotides)

match_with_db

match the resulting PWMs with motif databases using pssm_match. Note that the closest match is returned, even if it is not similar enough in absolute terms.

motif_dataset

a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example HOMER_motifs, JASPAR_motifs or all_motif_datasets(). By default all_motif_datasets() would be used.

multi_kmers

if TRUE, different candidates of kmers would be regressed in order to find the best seed according to final_metric.

final_metric

metric to use in order to choose the best motif. One of 'ks' or 'r2'. Note that unlike score_metric which is used in the regression itself, this metric is used only for choosing the best motif out of all the runs on the sampled dataset.

kmer_length

a vector of kmer lengths to screen in order to find the best seed motif.

max_cands

maximum number of kmer candidates to try.

motif_num

Number of motifs to infer. When motif_num > 1, the function would run motif_num times, each time on the residuals of a linear model of all the previous runs (see smooth_k parameter). The best motif is then returned, while all the others are stored at 'models' in the return value.

smooth_k

k for smoothing the predictions of each model in order to compute the residuals when motif_num > 1. The residuals are computed as response - running mean of size 'k' of the current model.

min_kmer_cor

minimal correlation between the kmer and the response in order to use it as a seed.

is_train

a boolean vector that determine which subset of sequences to use when screening

min_gap,max_gap

the length of a gap to be considered in the pattern. Only one gap, of length min_gap:max_gap, is being used, and is located anywhere in the motif. Note that this greatly expand the search space (and increase multiple testing severely).

sample_frac

fraction of the dataset to sample. When response is categorical (0 and 1), the sampling would be stratified by the category, i.e. sample_frac can be a vector of length 2 with the fraction of 0 and 1 responses to sample respectively. If NULL - the default would be 0.1 for continuous variables, and for binary variables - the number of 0 responses would be equal to sample_ratio times the number of 1 responses.

sample_idxs

indices of the sequences to use. If NULL, the indices would be sampled using sample_frac.

sample_ratio

ratio between the '1' category and the '0' category in the sampled dataset. Relevant only when sample_frac is NULL.

Value

a list with the following elements:

cv_models: a list of models, one for each fold.

cv_pred: a vector of predictions for each sequence.

score: score of the model on the cross-validated predictions.

cv_scores: a vector of scores for each fold.

folds: a vector with the fold number for each sequence.

full_model: The full model (without cross-validation), if add_full_model is TRUE.

Examples

if (FALSE) {
res <- regress_pwm.cv(
    cluster_sequences_example, cluster_mat_example[, 1],
    nfolds = 5, use_sample = TRUE, sample_frac = c(0.1, 1)
)
res$score
res$cv_scores

plot(
    res$cv_pred,
    res$full_model$pred,
    xlab = "CV predictions", ylab = "Full model predictions", cex = 0.1
)
plot_regression_prediction_binary(res$cv_pred, cluster_mat_example[, 1])
plot_regression_prediction_binary(res$full_model$pred, cluster_mat_example[, 1])

# without sampling
res <- regress_pwm.cv(
    cluster_sequences_example, cluster_mat_example[, 1],
    nfolds = 5, use_sample = FALSE
)
res$score
res$cv_scores
plot(res$cv_pred,
    res$full_model$pred,
    xlab = "CV predictions", ylab = "Full model predictions", cex = 0.1
)
}