Run PWM regression on a sample of the data

The optimization would be performed with a sampled dataset of size sample_frac, or explicit sampled indices sample_idxs. Note that multi_kmers is TRUE by default.

Usage

regress_pwm.sample(
  sequences,
  response,
  bidirect = TRUE,
  spat_min = 0,
  spat_max = NULL,
  include_response = TRUE,
  motif_num = 1,
  multi_kmers = TRUE,
  sample_frac = NULL,
  sample_idxs = NULL,
  sample_ratio = 1,
  parallel = getOption("prego.parallel", TRUE),
  match_with_db = FALSE,
  screen_db = FALSE,
  motif_dataset = all_motif_datasets(),
  seed = 60427,
  final_metric = NULL,
  unif_prior = 0.05,
  ...
)

Arguments

sequences

A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through toupper)

response

A matrix of response variables - number of rows should equal the number of sequences

bidirect

is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.

spat_min

start of the spatial model from the beginning of the sequence (in bp)

spat_max

end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model would end at the end of the sequence.

include_response

include the response in the resulting list (default: TRUE)

motif_num

Number of motifs to infer. When motif_num > 1, the function would run motif_num times, each time on the residuals of a linear model of all the previous runs (see smooth_k parameter). The best motif is then returned, while all the others are stored at 'models' in the return value.

multi_kmers

if TRUE, different candidates of kmers would be regressed in order to find the best seed according to final_metric.

sample_frac

fraction of the dataset to sample. When response is categorical (0 and 1), the sampling would be stratified by the category, i.e. sample_frac can be a vector of length 2 with the fraction of 0 and 1 responses to sample respectively. If NULL - the default would be 0.1 for continuous variables, and for binary variables - the number of 0 responses would be equal to sample_ratio times the number of 1 responses.

sample_idxs

indices of the sequences to use. If NULL, the indices would be sampled using sample_frac.

sample_ratio

ratio between the '1' category and the '0' category in the sampled dataset. Relevant only when sample_frac is NULL.

parallel

whether to run optimization in parallel. use set_parallel to set the number of cores to use.

match_with_db

match the resulting PWMs with motif databases using pssm_match. Note that the closest match is returned, even if it is not similar enough in absolute terms.

screen_db

Screen motif_dataset using screen_pwm and use the best motif as the initial motif. If TRUE, the following fields would be added to the return value: "db_motif", "db_motif_pred", "db_motif_pssm" and "db_motif_score".

motif_dataset

a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example HOMER_motifs, JASPAR_motifs or all_motif_datasets(). By default all_motif_datasets() would be used.

seed

random seed

final_metric

metric to use in order to choose the best motif. One of 'ks' or 'r2'. Note that unlike score_metric which is used in the regression itself, this metric is used only for choosing the best motif out of all the runs on the sampled dataset. If NULL - 'ks' would be used for binary response and 'r2' for continuous response.

unif_prior

uniform prior for nucleotide probabilities

...

Arguments passed on to regress_pwm, screen_kmers

motif

Initial motif to start the regression from. Can be either a string with a kmer where the character "*" indicates a wildcard or a data frame with a pre-computed PSSM (see the slot pssm in the return value of this function). If NULL - a K-mer screen would be performed in order to find the best kmer for initialization. If init_from_dataset is TRUE, the regression would be initialized from the PSSM of the best motif in the dataset.

init_from_dataset

initialize the regression from the PSSM of the best motif in motif_dataset, using final_metric as the metric. If TRUE, the motif parameter would be ignored. See screen_pwm for more details.

motif_length

Length of the seed motif. If the motif is shorter than this, it will be extended by wildcards (stars). Note that If the motif is longer than this, it will not be truncated.

score_metric

metric to use for optimizing the PWM. One of "r2" or "ks". When using "ks" the response variable should be a single vector of 0 and 1.

spat_bin

size of the spatial bin (in bp).

spat_model

a previously computed spatial model (see spat) in the return value of this function. This can only be used when motif is a previously computed PSSM.

improve_epsilon

minimum improve in the objective function to continue the optimization

min_nuc_prob

minimum nucleotide probability in every iteration

verbose

show verbose messages.

consensus_single_thresh,consensus_double_thresh

thresholds for the consensus sequence calculation (single and double nucleotides)

kmer_length

a vector of kmer lengths to screen in order to find the best seed motif.

max_cands

maximum number of kmer candidates to try.

smooth_k

k for smoothing the predictions of each model in order to compute the residuals when motif_num > 1. The residuals are computed as response - running mean of size 'k' of the current model.

min_kmer_cor

minimal correlation between the kmer and the response in order to use it as a seed.

is_train

a boolean vector that determine which subset of sequences to use when screening

min_gap,max_gap

the length of a gap to be considered in the pattern. Only one gap, of length min_gap:max_gap, is being used, and is located anywhere in the motif. Note that this greatly expand the search space (and increase multiple testing severely).

min_cor

Only patterns for which the maximum correlation to one of the response variable is larger than min_cor will be reported

min_n

Only patterns for which the average number of occurrences in the sequences is larger than min_n will be reported

from_range

Sequences will be considered only from position from_range (default 0)

to_range

Sequences will be considered only up to position to_range (default NULL - using the length of the sequences)

return_mat

Return a matrix of patterns and their correlation to the response variables instead of a data frame. (default: FALSE)

Value

a list with the following elements:

pssm: data frame with the pssm matrix with the inferred motif, where rows are positions and columns are nucleotides.

spat: a data frame with the inferred spatial model, with the spatial factor for each bin. The bins are defined such that the first bin starts at spat_min and the last bin ends at spat_max, with a bin size of spat_bin.

pred: a vector with the predicted pwm for each sequence.

consensus: Consensus sequence based on the PSSM.

response: The response matrix. If include_response is FALSE, the response matrix is not included in the list.

r2: \(r^2\) of the prediction with respect to the each response variable.

ks: If response is binary, Kolmogorov-Smirnov test results of the predictions where the response was 1 vs the predictions where the response was 0.

seed_motif: The seed motif that started the regression.

kmers: The k-mers that were screened in order to find the best seed motif (if motif was NULL).

sample_idxs: The indices of the sequences that were used for the regression (only for regress_pwm.sample).

When match_with_db is TRUE, the following additional elements are returned:

motif_db: The motif database that the most similar to the resulting PSSM.

db_match_dist: The distance between the resulting PSSM and the closest match in the motif database (KL divergence).

db_match_pssm: The PSSM of the closest match in the motif database.

db_match_pred: The predicted PWM of the closest match in the motif database.

db_match_r2: The \(r^2\) of the predicted PWM of the closest match in the motif database and the response

db_match_ks: If response is binary, the Kolmogorov-Smirnov test results of the predicted PWM of the closest match in the motif database where the response was 1 vs the predictions where the response was 0.

When screen_db is TRUE, the following additional elements are returned:

db_motif: The best motif from the motif database.

db_motif_pred: The predicted PWM of the best motif from the motif database.

db_motif_pssm: The PSSM of the best motif from the motif database.

db_motif_score: The score of the best motif from the motif database.

When n_motifs is greater than 1, the following additional elements are returned:

models: A list (as above) of each inferred model

multi_stats: A data frame with the following columns: model, score (KS for binary, r^2 otherwise), comb_score (score for the combined linear model for models 1:i) and additional statistics per model

Examples

if (FALSE) {
res <- regress_pwm.sample(cluster_sequences_example, cluster_mat_example[, 1], final_metric = "ks", screen_db = TRUE)
res$pssm
res$spat
head(res$pred)

plot_regression_qc(res)
}