Run a 2-phase PWM regression

The first phase of the optimization would be performed with a sampled dataset of size two_phase_sample_frac where different candidates of kmers would be regressed in order to find the best seed. Thenm the optimization would be performed on the full dataset while initializing the motif from the sampled dataset. You can also give explicit indices of the sequences to use in the first phase using first_phase_idxs.

Usage

regress_pwm_two_phase(
  sequences,
  response,
  motif_length = 15,
  score_metric = "r2",
  bidirect = TRUE,
  spat_min = 0,
  spat_max = NULL,
  spat_bin = 50,
  improve_epsilon = 0.0001,
  min_nuc_prob = 0.001,
  unif_prior = 0.05,
  is_train = NULL,
  include_response = TRUE,
  seed = 60427,
  verbose = FALSE,
  kmer_length = 5:8,
  min_gap = 0,
  max_gap = 1,
  min_kmer_cor = 0.1,
  two_phase_sample_frac = 0.1,
  first_phase_idxs = NULL,
  first_phase_metric = "r2",
  ...
)

Arguments

sequences

A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through toupper)

response

A matrix of response variables - number of rows should equal the number of sequences

motif_length

Length of the seed motif. If the motif is shorter than this, it will be extended by wildcards (stars). Note that If the motif is longer than this, it will not be truncated.

score_metric

metric to use for optimizing the PWM. One of "r2" or "ks". When using "ks" the response variable should be a single vector of 0 and 1.

bidirect

is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.

spat_min

start of the spatial model from the beginning of the sequence (in bp)

spat_max

end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model would end at the end of the sequence.

spat_bin

size of the spatial bin (in bp).

improve_epsilon

minimum improve in the objective function to continue the optimization

min_nuc_prob

minimum nucleotide probability in every iteration

unif_prior

uniform prior for nucleotide probabilities

is_train

a boolean vector that determine which subset of sequences to use when screening

include_response

include the response in the resulting list (default: TRUE)

seed

random seed

verbose

verbosity of the second phase of the optimization.

kmer_length

a vector of kmer lengths to screen in order to find the best seed motif.

two_phase_sample_frac

fraction of the dataset to sample for the first phase of the optimization (default: 0.1). When response is categorical (0 and 1), the sampling would be stratified by the category, i.e. two_phase_sample_frac can be a vector of length 2 with the fraction of 0 and 1 responses to sample respectively.

first_phase_idxs

indices of the sequences to use in the first phase of the optimization. If NULL, the indices would be sampled using two_phase_sample_frac.

first_phase_metric

metric to use in order to choose the best motif in the first phase of the optimization. One of 'ks' or 'r2'. Note that unlike score_metric which is used in the regression itself, this metric is used only for choosing the best motif in the first phase of the optimization out of all the runs on the sampled dataset.

Value

a list with the following elements:

pssm: data frame with the pssm matrix with the inferred motif, where rows are positions and columns are nucleotides.

spat: a data frame with the inferred spatial model, with the spatial factor for each bin.

pred: a vector with the predicted pwm for each sequence.

response: The response matrix. If include_response is FALSE, the response matrix is not included in the list.

r2: \(r^2\) of the prediction with respect to the each response variable.

ks: If response is binary, Kolmogorov-Smirnov test results of the predictions where the response was 1 vs the predictions where the response was 0.

seed_motif: The seed motif that started the regression.

kmers: The k-mers that were screened in order to find the best seed motif (if motif was NULL).

Examples

res <- regress_pwm_two_phase(cluster_sequences_example, cluster_mat_example[, 1], two_phase_sample_frac = c(0.1, 1))
#> Error in regress_pwm_two_phase(cluster_sequences_example, cluster_mat_example[,     1], two_phase_sample_frac = c(0.1, 1)): could not find function "regress_pwm_two_phase"
res$pssm
#> Error in eval(expr, envir, enclos): object 'res' not found
res$spat
#> Error in eval(expr, envir, enclos): object 'res' not found
head(res$pred)
#> Error in head(res$pred): object 'res' not found

plot_regression_qc(res)
#> Error in dplyr::select(., -1): object 'res' not found